+++ FundamentalsCore/src/com/johnymuffin/beta/fundamentals/commands/CommandBalance.java
@@ -3,6 +3,7 @@ package com.johnymuffin.beta.fundamentals.commands;
 import com.johnymuffin.beta.fundamentals.api.EconomyAPI;
 import com.johnymuffin.beta.fundamentals.api.FundamentalsAPI;
 import com.johnymuffin.beta.fundamentals.settings.FundamentalsLanguage;
+import com.johnymuffin.beta.fundamentals.util.Utils;
 import com.projectposeidon.api.PoseidonUUID;
 import com.projectposeidon.api.UUIDType;
 import org.bukkit.command.Command;
@@ -57,17 +58,11 @@ public class CommandBalance implements CommandExecutor {
         }
         return true;
     }
-    public static double round(double value, int places) {
-        if (places < 0) throw new IllegalArgumentException();
-    
-        BigDecimal bd = BigDecimal.valueOf(value);
-        bd = bd.setScale(places, RoundingMode.HALF_UP);
-        return bd.doubleValue();
-    }
+
     private void printBalance(String playerName, CommandSender commandSender, EconomyAPI.BalanceWrapper balanceWrapper) {
         if (balanceWrapper.getEconomyResult() == EconomyAPI.EconomyResult.successful) {
             String message = FundamentalsLanguage.getInstance().getMessage("balance_successful");
-            to_return = round(balanceWrapper.getBalance(), 2)
+            to_return = Utils.round(balanceWrapper.getBalance(), 2);
             message = message.replace("%var1%", String.valueOf((double) to_return);
             commandSender.sendMessage(message);
         } else if (balanceWrapper.getEconomyResult() == EconomyAPI.EconomyResult.userNotKnown) {
