+++ FundamentalsCore/src/com/johnymuffin/beta/fundamentals/util/Utils.java
@@ -21,6 +21,13 @@ import java.util.regex.Pattern;
 
 public class Utils {
 
+    public static double round(double value, int places) {
+        if (places < 0) throw new IllegalArgumentException();
+    
+        BigDecimal bd = BigDecimal.valueOf(value);
+        bd = bd.setScale(places, RoundingMode.HALF_UP);
+        return bd.doubleValue();
+    }
 
     public static Player getPlayerFromString(String name) {
         List<Player> players = Bukkit.matchPlayer(name);
